+++ src/core/Config.js
@@ -1,0 +1,15 @@
+export class Config {
+    static CHAT_MAX_LENGTH = 200;
+    static CHAT_DISPLAY_TIME = 8000;
+    static CHAT_FADE_TIME = 1000;
+    static EMOTE_DISPLAY_TIME = 3000;
+    static CAMERA_MIN_DISTANCE = 2;
+    static CAMERA_MAX_DISTANCE = 8;
+    static PLAYER_MOVE_SPEED = 0.05;
+    static PLAYER_JUMP_SPEED = 0.18;
+    static PLAYER_GRAVITY = -0.005;
+    
+    constructor() {
+        // Configuraciones dinámicas si son necesarias
+    }
+}
